Refactor UI exports for React Fast Refresh compliance
