+++ src/components/mode-toggle.tsx
@@ -2,7 +2,7 @@ import { Moon, Sun } from "lucide-react"
 
 import { Button } from "@/components/ui/button"
 import {    DropdownMenu,    DropdownMenuContent,    DropdownMenuItem,    DropdownMenuTrigger,} from "@/components/ui/dropdown-menu"
-import { useTheme } from "@/components/theme-provider"
+import { useTheme } from "@/components/theme-provider-context"
 
 export function ModeToggle() {
     const { setTheme } = useTheme()
