+++ src/components/theme-provider-context.ts
@@ -1,0 +1,25 @@
+import { createContext, useContext } from "react"
+
+type Theme = "dark" | "light" | "system"
+
+type ThemeProviderState = {
+  theme: Theme
+  setTheme: (theme: Theme) => void
+}
+
+const ThemeProviderContext = createContext<ThemeProviderState | undefined>(
+  undefined
+)
+
+function useTheme() {
+  const context = useContext(ThemeProviderContext)
+
+  if (!context) {
+    throw new Error("useTheme must be used within a ThemeProvider")
+  }
+
+  return context
+}
+
+export { ThemeProviderContext, useTheme }
+export type { Theme, ThemeProviderState }
