+++ src/components/ui/toggle-group.tsx
@@ -2,8 +2,8 @@ import * as React from "react"
 import * as ToggleGroupPrimitive from "@radix-ui/react-toggle-group"
 import { type VariantProps } from "class-variance-authority"
 
+import { toggleVariants } from "@/components/ui/toggle-variants"
 import { cn } from "@/lib/utils"
-import { toggleVariants } from "@/components/ui/toggle"
 
 const ToggleGroupContext = React.createContext<
   VariantProps<typeof toggleVariants>
